% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseMarginalFPCA.R
\name{DenseMarginalFPCA}
\alias{DenseMarginalFPCA}
\title{Marginal Functional Principal Component Analysis for dense functional data.}
\usage{
DenseMarginalFPCA(
  X.age.year,
  n,
  num.years,
  num.ages,
  fpca.op1 = list(),
  fpca.op2 = list(),
  pc.j = NULL,
  pc.k = NULL
)
}
\arguments{
\item{X.age.year}{An n by (num.years*num.ages) input data matrix, such that the ith row of the matrix gives the observed values for the ith individual. The values in each row are sorted first by years (dimension 1) and then by ages (dimension 2) within each year.}

\item{n}{The sample size.}

\item{num.years}{Dimension 1}

\item{num.ages}{Dimension 2}

\item{fpca.op1}{A list of options control parameters specified by list(name=value) for FPCA in direction of (age) dimension 2; check fdapace::FPCA for available control options and default settings.}

\item{fpca.op2}{A list of options control parameters specified by list(name=value) for FPCA in direction of (year) dimension 1; check fdapace::FPCA for available control options and default settings.}

\item{pc.j}{A scalar denoting the maximum number of components to consider for marginal FPCA in direction of (age) dimension 2; default: chosen by FVE if NULL}

\item{pc.k}{A vector of length pc.j denoting the maximum number of components to consider for nested FPCA in direction of (year) dimension 1; default: chosen by FVE if NULL}
}
\value{
A list containing the following fields:
\item{Xest}{An n by (num.years*num.ages) estimated matrix, based on the fitted Marginal FPCA model. The ith row of the matrix gives the observed values for the ith individual. The values in each row are sorted first by years (dimension 1) and then by ages (dimension 2) within each year.}
\item{mu}{An num.ages by num.years matrix containing the bivariate mean function estimate.}
\item{pc.j}{A scalar denoting the selected number of components for FPCA in direction of (age) dimension 2.}
\item{pc.k}{A vector of length pc.j, denoting the selected number of components for FPCA in direction of (year) dimension 1.}
\item{scores}{An n by sum(pc.k) matrix of the estimated scores, such that the ith row of the matrix comprises estimated scores chi_{1,1},chi_{1,2},...chi_{1,pc.k[1]},chi_{2,1},chi_{2,2},...,chi_{2,pc.k[2]},...,chi_{pc.j,1},chi_{pc.j,2},...,chi_{pc.j,pc.k[j]} for the ith individual.}
\item{res.psi}{A list containing the FPCA output for FPCA in direction of (age) dimension 2.}
\item{res.phi}{A list containing the FPCA output for FPCA in direction of (year) dimension 1.}
\item{eig}{An (num.years*num.ages) by sum(pc.k) matrix of the estimated product eigen functions. The rows are sorted first by years (dimension 1) and then by ages (dimension 2) within each year. The columns are sorted as in the scores.}
\item{psi}{An num.ages by pc.j matrix containing the estimated eigenfunctions for FPCA in direction of (age) dimension 2.}
\item{phi}{An num.years by sum(pc.k) matrix, containing the estimated eigenfunctions for FPCA in direction of (year) dimension 1.}
\item{FVE}{A vector of length sum(pc.k), indicating the fraction of total variance explained by each product function, with corresponding 'FVEthreshold'.}
\item{VarOrdered}{Variance explained by each term. The terms are ordered by var(chi_{jk}). One can select the best model by truncating at a desired level of FVE, and use names(VarOrdered) to see the corresponding model terms.}

 @examples
n <- 100 ### sample size
num.ages <- 20 ### dimension 2
num.years <- 15 ### dimension 1
dense_grid <- seq(0,1,length=N)
Lt <- list()
Ly <- list()
for (i in 1:n) {
  Lt[[i]] <- dense_grid ### dense time grid
  y_temp <- matrix(0,num.ages,num.years)
  for (s in 1:num.ages) {
    for (t in 1:num.years) {
      y_temp[s,t] <- y_temp[s,t]+cos(Lt[[i]][t])+rnorm(1,0,0.5)
    }
  }
  Ly[[i]] <- y_temp ### dense functional data
}
X.age.year <- matrix(0,n,num.years*num.ages)
for (i in 1:n) {
  X.age.year[i,] <- as.vector(Ly[[i]]) ### data matrix
}
res<-DenseMarginalFPCA(X.age.year, n, 15, 20, fpca.op1=NULL, fpca.op2=NULL, pc.j = NULL, pc.k = NULL)
# Basic output
res$Xest
res$mu
res$pc.j
res$pc.k
res$scores
res$res.psi
res$psi
res$FVE
res$VarOrdered
#Additional scores
fpca_psi <- res$res.psi
xi_i_j <- fpca_psi$xiEst
str(fpca_psi$xiEst)
#xi{1,}(t) scores for 1st individual
xi_i_j[1:num.years,]
#xi{1,}(t) scores for 2nd individual
xi_i_j[(num.years+1):(2*num.years),]
}
\description{
Note: The code works for dense functional data observed on a regular grid, missing values are allowed, written by Kehui Chen 10/09/2015, based on the original code by Pedro Delicado.
}
\details{
The code works for dense functional data (with missing values), with density in both the direction of (age) dimension 2 and (year) dimension 1.
}
\references{
\itemize{
\item \cite{Chen, K., Delicado, P., & M端ller, H. G. (2017). Modelling function-valued stochastic processes, with applications to fertility dynamics. Journal of the Royal Statistical Society Series B: Statistical Methodology, 79(1), 177-196.}
\item \cite{Chen, K., & M端ller, H. G. (2012). Modeling repeated functional observations. Journal of the American Statistical Association, 107(500), 1599-1609.}
\item \cite{Hall, P.,  M端ller, H.G. and Wang, J.L. (2006). Properties of principal component methods for functional and longitudinal data analysis. Annals of Statistics, 34(3), 1493-1517.}
\item \cite{Yao, F., M端ller, H. G., & Wang, J. L. (2005). Functional data analysis for sparse longitudinal data. Journal of the American statistical association, 100(470), 577-590.}
}
}
